pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astk"
version = "0.1.0"
description = "AgentSprint TestKit - A CI-first behavioural-coverage and regression-gating framework for AI agents"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    { name = "ASTK Team" },
]
dependencies = [
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "deepeval>=0.20.0",
    "garak>=0.9.0",
    "PyYAML>=6.0.1",
    "docker>=7.0.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.1",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.1",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100 